Moved get_datalink_channel, get_interface, lookup_addr, sigwinch, and create_write_to_stdout
