+++ src/os/linux.rs
@@ -1,38 +1,18 @@
-use ::pnet::datalink::Channel::Ethernet;
-use ::pnet::datalink::DataLinkReceiver;
-use ::pnet::datalink::{self, Config, NetworkInterface};
-use ::std::io::{self, Write};
-
 use ::std::collections::HashMap;
-use ::std::net::IpAddr;
-use ::std::time;
 
 use ::procfs::FDTarget;
-use signal_hook::iterator::Signals;
 
 use crate::network::{Connection, Protocol};
 use crate::OsInputOutput;
 
-use crate::os::shared::{KeyboardEvents};
-
-
-fn get_datalink_channel(
-    interface: &NetworkInterface,
-) -> Result<Box<dyn DataLinkReceiver>, failure::Error> {
-    let mut config = Config::default();
-    config.read_timeout = Some(time::Duration::new(0, 1));
-    match datalink::channel(interface, config) {
-        Ok(Ethernet(_tx, rx)) => Ok(rx),
-        Ok(_) => failure::bail!("Unknown interface type"),
-        Err(e) => failure::bail!("Failed to listen to network interface: {}", e),
-    }
-}
-
-fn get_interface(interface_name: &str) -> Option<NetworkInterface> {
-    datalink::interfaces()
-        .into_iter()
-        .find(|iface| iface.name == interface_name)
-}
+use crate::os::shared::{
+    KeyboardEvents,
+    get_datalink_channel,
+    get_interface,
+    lookup_addr,
+    sigwinch,
+    create_write_to_stdout,
+};
 
 fn get_open_sockets() -> HashMap<Connection, String> {
     let mut open_sockets = HashMap::new();
@@ -74,37 +54,7 @@ fn get_open_sockets() -> HashMap<Connection, String> {
     open_sockets
 }
 
-fn lookup_addr(ip: &IpAddr) -> Option<String> {
-    ::dns_lookup::lookup_addr(ip).ok()
-}
 
-fn sigwinch() -> (Box<dyn Fn(Box<dyn Fn()>) + Send>, Box<dyn Fn() + Send>) {
-    let signals = Signals::new(&[signal_hook::SIGWINCH]).unwrap();
-    let on_winch = {
-        let signals = signals.clone();
-        move |cb: Box<dyn Fn()>| {
-            for signal in signals.forever() {
-                match signal {
-                    signal_hook::SIGWINCH => cb(),
-                    _ => unreachable!(),
-                }
-            }
-        }
-    };
-    let cleanup = move || {
-        signals.close();
-    };
-    (Box::new(on_winch), Box::new(cleanup))
-}
-
-pub fn create_write_to_stdout() -> Box<dyn FnMut(String) + Send> {
-    Box::new({
-        let mut stdout = io::stdout();
-        move |output: String| {
-            writeln!(stdout, "{}", output).unwrap();
-        }
-    })
-}
 
 pub fn get_input(interface_name: &str) -> Result<OsInputOutput, failure::Error> {
     let keyboard_events = Box::new(KeyboardEvents);
